Restored from the previous (failed) ControlsRefactor branch:
Major refactoring for controls to eliminate redundancy.
Click/touch handlers are executed in this order: first for widgets, then for game objects and last for the background. If multiple listeners get triggered in a location, only most precedent of them get executed.

Will be merged to trunk after testing (with real WP7 and all the examples) is complete.